Affichage des buttons de pagination dans la page

+++ index.js
@@ -1,5 +1,6 @@
 const btnCreerTache = document.querySelector("#creerTache");
 const createTacheSection = document.querySelector("section.main_action");
+const form = document.querySelector("form");
 const btnAnnulerTache = document.querySelector("#annulerTache");
 const btnAjouterTache = document.querySelector("#ajouterTache");
 const p = document.querySelector("section.main_list > p");
@@ -34,9 +35,11 @@ function handleTacheEvents() {
     const btnCompleterTacheList = document.querySelectorAll("#completerTache");
 
     btnSupprimerTacheList.forEach(btnSupprimer => btnSupprimer.addEventListener('click', (e) => {
-        const tacheID = e.target.parentElement.parentElement.parentElement.getAttribute('key');
-        supprimerTache(parseInt(tacheID));
-        isTacheAjouteeState = true;
+        if (confirm(`Voulez-vous réellement supprimer cette tâche ?`)) {
+            const tacheID = e.target.parentElement.parentElement.parentElement.getAttribute('key');
+            supprimerTache(parseInt(tacheID));
+            isTacheAjouteeState = true;
+        }
     }));
 
     btnCompleterTacheList.forEach(btnCompleter => btnCompleter.addEventListener('click',
@@ -52,20 +55,12 @@ function handleTacheEvents() {
 function renderTaches() {
     console.log("Rendering taches...");
     if (taches.length > 0) {
-        tachesList.innerHTML = '';
         p.style.display = 'none';
         try {
-            taches.map(({ tacheID, nom, dateEcheance, categorie, priorite, estTerminee, rappels, _dateCreation }) => {
-                const tacheObj = creerObjetTache();
-                tacheObj.defineTache(tacheID, nom, dateEcheance, priorite, estTerminee, categorie, rappels);
-                tacheObj.setDateCreation(_dateCreation);
-                tachesList.innerHTML += TacheCardComponent(tacheObj);
-            });
+            //Affiche la pagination
+            mainListSection.innerHTML = renderPagination(taches);
 
             handleTacheEvents();
-
-            //Affiche la pagination
-            // mainListSection.innerHTML += renderPagination();
         } catch (error) {
             p.style.display = 'block';
             console.error(error);
@@ -77,6 +72,7 @@ function renderTaches() {
 
 // Events
 btnCreerTache.addEventListener("click", () => {
+    form.reset();
     createTacheSection.style.display = "block";
 });
 
@@ -88,6 +84,7 @@ btnAjouterTache.addEventListener("click", (e) => {
     e.preventDefault();
     const tacheID = taches.length == 0 ? taches.length + 1 : parseInt(taches[taches.length - 1].tacheID) + 1;
     const selectedRappel = document.querySelector('input[name="rappels"]:checked').value;
+    const tempsRappel = [`${rappel.value}${selectedRappel}`];
 
     if (new Date(dateEcheance.value).getTime() < new Date().getTime()) {
         alert(`La date d'échéance n'est pas valide`);
@@ -97,6 +94,6 @@ btnAjouterTache.addEventListener("click", (e) => {
 
     createTacheSection.style.display = "none";
 
-    creerTache(tacheID, tacheNom.value, dateEcheance.value, priorité.value, false, categorie.value, `${rappel.value}${selectedRappel}`);
+    creerTache(tacheID, tacheNom.value, dateEcheance.value, priorité.value, false, categorie.value, tempsRappel);
     isTacheAjouteeState = true;
 });
